Front end folder refactor
- Pages now have their own folder under the pages folder
- All components relevant to a specific page live under their respective folders under the pages folder
- AnimatedRoutes renamed to Loading for more consistency. (could be confused with routing in google maps)
- Footer component placed in the main component directory with the rest of the general components
- Added google MUI loading circle, can choose which one to remove later
