+++ tecttitans-app/src/data/getESGScore.js
