+++ tecttitans-app/src/pages/Home/HeroSection.test.js
